Move scroll hook to hooks folder.
